requirements.txt
streamlit
Pillow
opencv-python
numpy
pytest
razorpay
python-dotenv
scikit-learn
reportlab